Attempt to enforce HTTP/1.1 only
Noting that there is a mismatch between negotiating the extension
at the TLS layer and processing HTTP 1.x requests that state their
HTTP version inline, attempt to actually enforce the otherwise
stated requirement to only use this with HTTP/1.1.

Close the connection in this case, akin to a mismatched header field,
since the state of the connection is probably not reliable for much
anymore.